refactor: seperate class to ticket generator
